+++ lib/rails_memory_profiler/engine.rb
@@ -2,5 +2,9 @@ module RailsMemoryProfiler
   class Engine < ::Rails::Engine
     isolate_namespace RailsMemoryProfiler
     config.generators.api_only = true
+
+    initializer "rails_memory_profiler.middleware" do |app|
+      app.middleware.use(Middleware)
+    end
   end
 end